tests/requirements.txt
pytest==6.2.5
python-dateutil==2.8.2
paho-mqtt==1.2.3
pytest-sugar==0.9.4
pymongo==3.11.2
requests==2.28.1
Pillow==9.0.1